sem2/sp/blind-mitigation-main-weier/requirements.txt
bokeh==2.3.2
dill==0.3.4
imbalanced_learn==0.8.1
imblearn==0.0
matplotlib==3.4.3
numpy~=1.21
pandas==1.3.3
PyYAML==6.0
scikit_learn==1.0.1
seaborn==0.11.2
tensorflow==2.7.0
jupyterlab==3.2.2
markovify==0.9.4